added jpa and h2 in memory.

+++ src/main/java/com/stefanbanu/todowebservices/services/TodoHardcodedService.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Service
 public class TodoHardcodedService {
     private static List<Todo> todos = new ArrayList<>();
-    private static int idCounter = 0;
+    private static long idCounter = 0;
 
     static {
         todos.add(new Todo(++idCounter, "stefan", "learn spring boot", LocalDate.now(), false));
@@ -22,7 +22,7 @@ public class TodoHardcodedService {
         return todos;
     }
 
-    public Todo deleteById(long id) {
+    public Todo deleteById(Long id) {
         Todo todo = findById(id);
         if (todo == null) return null;
         if (todos.remove(todo)) return todo;
@@ -30,7 +30,7 @@ public class TodoHardcodedService {
         return null;
     }
 
-    public Todo findById(long id) {
+    public Todo findById(Long id) {
         for (Todo todo : todos) {
             if (todo.getId() == id) {
                 return todo;
